More precise output normalization (hopefully)
Hopefully, instead of dividing each component by a maximum value, we
normalize it, so that the magnitude is 1 and the individual components
are in the range of [-1, 1]. In theory, this should change nothing at
worse and make faster movements? at best. However, I am almost certain
that I do not know what I am doing.

IF THIS CHANGE CAUSES THE ROBOT TO SHAKE VIOLENTLY, NOTIFY ME ASAP.